+++ CHANGELOG.md
@@ -16,7 +16,7 @@ fixed since the last alpha release.
     - Pulse
     - PulseNote
     - PulseUser
-- Objects now return an `InvalidArgumentException` when given null to the constructor
+- Objects now throw an `InvalidArgumentException` when given null to the constructor
 - Pulse column objects no longer attempt to create null objects and instead return null when asked for the value
 - Introduce `PulseColumnStatusValue::getHexColor()` to get hex values of the respective statuses
 - `PulseColumnPersonValue::updateValue` now supports passing a PulseUser object in addition to just an int